chore: Add localization files and update dependencies
