+++ lib/l10n/app_en.arb
@@ -1,0 +1,3 @@
+{
+  "selectImageFolder": "Select Images Folder"
+}
