+++ lib/l10n/app_es.arb
@@ -1,0 +1,3 @@
+{
+    "selectImageFolder": "Selecciona la carpeta de imágenes"
+}
