+++ lib/theme/thermal_theme.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ThermalTheme {
+  static ThemeData get theme {
+    return ThemeData(
+      colorScheme: const ColorScheme.highContrastLight(),
+      useMaterial3: true,
+    );
+  }
+}
