+++ pubspec.yaml
@@ -30,6 +30,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter
+  intl: any
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -39,6 +42,7 @@ dependencies:
   file_picker: ^8.0.6
   url_launcher: ^6.3.0
   image: ^4.2.0
+  path: ^1.9.0
 
 platforms:
   windows:
@@ -60,6 +64,8 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
+
+  generate: true
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
